desktop: the Windows service, and a Windows build that actually compiles
The service was declared and never written. It is here now, split the way
the rest of the workspace is: every decision in `machine.rs`, which
compiles and is tested on any host, and only the SCM, the named pipe and
the restart in the Windows half.

  * install/uninstall register and remove the service cleanly, stopping
    it first so `delete` takes effect now rather than at the next reboot
    — a leftover privileged service is what an administrator will not
    forgive;
  * the pipe is created with the reviewed SDDL, rejects remote clients,
    bounds its instances, and drops a client that falls silent;
  * a restart is refused for a session the service was never told about
    and rate-limited against a restart loop;
  * the service does not start the UI. Windows removed interactive
    services; the installer registers the tray application at sign-in
    instead, and the doc comment claiming otherwise is corrected.

Cross-checking the workspace against x86_64-pc-windows-gnu found that the
Windows platform layer had never compiled: eight missing `windows` crate
features and four API mismatches (BOOL moved to windows-core,
MONITORINFOF_PRIMARY to WindowsAndMessaging, OpenInputDesktop to
StationsAndDesktops, CRYPTOAPI_BLOB renamed, D3D11CreateDevice's HMODULE,
SendInput's unsigned mouseData). All fixed; both targets now check clean
with no warnings.

Also gives the agent a machine-wide configuration file, written
atomically, so the service and the tray application cannot end up
pointing at different deployments.

+++ desktop/src-tauri/src/platform/windows/input.rs
@@ -24,8 +24,8 @@
 
 use std::sync::Mutex;
 
-use remote_device::{InputProvider, PlatformError, PlatformResult};
-use remote_protocol::{Key, KeyEvent, MouseButton, MouseEvent, PointerPosition, ScrollEvent};
+use remote_device::{InputProvider, PlatformResult};
+use remote_protocol::{Key, KeyEvent, MouseEvent, PointerPosition, ScrollEvent};
 
 /// One notch of a wheel, as Windows counts it.
 pub const WHEEL_DELTA: i32 = 120;
@@ -326,6 +326,11 @@ mod imp {
 
     /// Where a normalised position on one monitor lands on the virtual desktop.
     fn resolve(monitor_id: u32, position: PointerPosition) -> PlatformResult<(i32, i32)> {
+        // The trait has to be in scope for `monitors()`; it is the same
+        // provider the capture side uses, so the coordinates a click is
+        // resolved against are the ones the viewer was looking at.
+        use remote_device::ScreenCaptureProvider;
+
         let layout = super::super::capture::WindowsCapture::new().monitors()?;
 
         let monitor = layout
@@ -366,7 +371,9 @@ mod imp {
                 mi: MOUSEINPUT {
                     dx: x,
                     dy: y,
-                    mouseData: data,
+                    // `mouseData` is unsigned; a negative scroll notch is the
+                    // same bit pattern, which is what Windows reads it as.
+                    mouseData: data as u32,
                     dwFlags: flags,
                     time: 0,
                     dwExtraInfo: 0,
@@ -531,6 +538,9 @@ mod imp {
 #[cfg(test)]
 mod tests {
     use super::*;
+    // Used only by the no-implementation test below, which is the one that
+    // does not compile on Windows.
+    #[cfg(not(target_os = "windows"))]
     use remote_protocol::Modifiers;
 
     fn desktop() -> VirtualDesktop {
